refactor: move cleanup to client
