+++ alistral_cli/src/main.rs
@@ -1,10 +1,7 @@
 use clap::Parser;
 
 use crate::interface::tracing::init_tracer;
-use crate::models::client::AlistralCliClient;
-use database::cleanup::cleanup_database;
 use models::cli::Cli;
-use tracing::debug;
 
 pub(crate) mod api;
 pub(crate) mod database;
@@ -31,31 +28,14 @@ async fn main() -> ColEyre {
     let _worker_guard = init_tracer(&cli);
     //console_subscriber::init();
 
-    if run_cli(cli).await {
-        post_run().await
+    match cli.run().await {
+        Result::Ok(clean) => {
+            if clean {
+                ALISTRAL_CLIENT.clean_up_mb_db().await;
+            }
+        }
+        Err(err) => err.panic(),
     }
 
     Ok(())
-}
-
-async fn run_cli(cli: Cli) -> bool {
-    match cli.run().await {
-        Result::Ok(val) => val,
-        Err(err) => err.panic(),
-    }
-}
-
-async fn post_run() {
-    let alistral_client = AlistralCliClient::create_or_fatal().await;
-    let conn = &mut *alistral_client
-        .musicbrainz_db
-        .get_raw_connection()
-        .await
-        .expect("Couldn't connect to the database");
-
-    debug!("Cleaning some old entries...");
-    cleanup_database(conn)
-        .await
-        .expect("Error while cleaning the database");
-    debug!("Done!");
 }
