+++ alistral_cli/src/tools/radio/overdue.rs
@@ -6,6 +6,7 @@ use futures::{StreamExt, stream};
 use interzic::models::playlist_stub::PlaylistStub;
 use tracing::info;
 
+use crate::ALISTRAL_CLIENT;
 use crate::datastructures::radio::collector::RadioCollector;
 use crate::datastructures::radio::filters::cooldown::cooldown_filter;
 use crate::datastructures::radio::filters::min_listens::min_listen_filter;
@@ -19,7 +20,6 @@ use crate::models::data_storage::DataStorage;
 use crate::models::error::ResultTEExt;
 use crate::tools::radio::convert_recordings;
 use crate::utils::data_file::DataFile;
-use crate::ALISTRAL_CLIENT;
 
 //TODO: Refactor Radios params into structs
 #[expect(clippy::too_many_arguments)]
